refactor: exclude unnecessary files from linting and testing
This commit removes several files from the list of excluded paths in .ansible-lint and .yamllint.yml. It also updates the molecule.yml file to install only necessary dependencies for testing. Additionally, it modifies converge.yml to disable fact gathering and privilege escalation, and changes the role name in that file. Finally, it simplifies ubuntu.yml by removing a redundant list.