feat: quality of life update
This commit adds an Ansible role for the Landscape client. It includes changes to default files, handlers, and tasks. The `ubuntu.yml` task file was renamed to `install.yml`. The `main.yml` task file now includes a block that installs Landscape on Ubuntu systems and configures it. A new `.vscode/settings.json` file was added with the path to the Python interpreter. Finally, a check was added to see if the client is already registered before registering it again.